Create README.md
As I begin my journey in Data science, I have seen it fit to share my the projects I'll be taking on. I am open to criticism because I am eager to learn. 

I'll start with with machine learning model to predict the housing prices of houses in Melbourne, using historical data that I'll split into training and test data.